feat: Add new commands: uptime, help and ping
This commit adds new commands to the bot: uptime, help and ping. The uptime command shows the bot's uptime, showing also the current shards' status. The help command shows all the commands that the bot has, divided by category. The ping command shows the bot's latency and the API latency.